server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai>=1.54.0
python-dotenv==1.0.1
yt-dlp>=2025.1.15
